LayoutTests/imported/w3c/web-platform-tests/html/semantics/embedded-content/the-embed-element/embed-network-error.sub-expected.txt

CONSOLE MESSAGE: Refused to display 'http://web-platform.test:8800/html/semantics/embedded-content/resources/not-embeddable.html' in a frame because it set 'X-Frame-Options' to 'deny'.
CONSOLE MESSAGE: Refused to display 'http://web-platform.test:8800/html/semantics/embedded-content/resources/not-embeddable.html' in a frame because it set 'X-Frame-Options' to 'deny'.


Harness Error (TIMEOUT), message = null

TIMEOUT new embed: nonexistent host Test timed out
PASS new embed: X-Frame-Options prevents embedding
TIMEOUT navigating an existing embed: nonexistent host Test timed out
PASS navigating an existing embed: X-Frame-Options prevents embedding
